add mechanism to invoke logging of statistics
